TalkoTestProgram finished
This is the finished attempt at the Talko interview transformation test
program.

StrngUtilsTest.java - is the JUnit test of the functionality.
StringUtils.java - is the java code to achieve the purposed
functionality
StringNode - is a structure that needed creating to achieve the
requirements.
dictionary.txt - is in a folder up above called assets (sort of stolen
from the Android naming convention) which stores the dictionary list of
five letter words.